requirements.txt
attrs==19.1.0
colorama==0.4.1
PyQt5==5.13.0
PyHTML==1.3.1
PyGObject==3.32.2
PyYAML==5.1.2
ruamel.yaml==0.16.1
PyQtWebEngine==5.13.0
